added Custom methods & Handle deletebyId with ResponseEntity

+++ src/main/java/com/dailycodebuffer/springboot/tutorial_1/service/DepartmetServiceImpl.java
@@ -7,11 +7,13 @@ import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Objects;
+import java.util.Optional;
 
 @Service
-public class DepartmetServiceImpl implements DepartmentService{
+public class DepartmetServiceImpl implements DepartmentService {
     @Autowired
     private DepartmentRepository departmentRepository;
+
     @Override
     public Department saveDepartment(Department department) {
         return departmentRepository.save(department);
@@ -28,30 +30,40 @@ public class DepartmetServiceImpl implements DepartmentService{
     }
 
     @Override
-    public void deleteDepartmentById(Long departmentId) {
-        departmentRepository.deleteById(departmentId);
+    public int deleteDepartmentById(Long departmentId) {
+        Optional<Department> departmentOptional = departmentRepository.findById(departmentId);
+        if (departmentOptional.isPresent()) {
+            departmentRepository.deleteById(departmentId);
+            return 1;
+        } else return 0;
+
     }
 
     @Override
     public Department updateDepartment(Long departmentId, Department department) {
         Department db = departmentRepository.findById(departmentId).get();
 
-        if(Objects.nonNull(department.getDepartmentName())
+        if (Objects.nonNull(department.getDepartmentName())
                 && !"".equalsIgnoreCase(department.getDepartmentName())) {
             db.setDepartmentName(department.getDepartmentName());
         }
 
-        if(Objects.nonNull(department.getDepartmentAddress())
+        if (Objects.nonNull(department.getDepartmentAddress())
                 && !"".equalsIgnoreCase(department.getDepartmentAddress())) {
             db.setDepartmentAddress(department.getDepartmentAddress());
         }
 
-        if(Objects.nonNull(department.getDepartmentCode())
+        if (Objects.nonNull(department.getDepartmentCode())
                 && !"".equalsIgnoreCase(department.getDepartmentCode())) {
             db.setDepartmentCode(department.getDepartmentCode());
         }
 
         return departmentRepository.save(db);
+    }
+
+    @Override
+    public Department fetchDepartmentByName(String departmentName) {
+        return departmentRepository.findByDepartmentNameIgnoreCase(departmentName);
     }
